NotFoundPostException

+++ src/posts/post.repository.ts
@@ -2,6 +2,7 @@ import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { PostDo } from 'src/_schemas/post.do';
 import { UpdatePostDto } from './dto/update-post.dto';
+import { NotFoundPostException } from 'src/exceptions/not-found-exception.exception';
 
 export class PostsRepository {
   constructor(
@@ -10,6 +11,7 @@ export class PostsRepository {
   ) {}
   async createOne(post: PostDo): Promise<PostDo> {
     const createOne = await this.postModel.create(post);
+
     return createOne;
   }
   async findAll(): Promise<any> {
@@ -18,6 +20,10 @@ export class PostsRepository {
   }
   async findOne(id: string): Promise<any> {
     const findOne = await this.postModel.findById(id);
+    if (!findOne) {
+      throw new NotFoundPostException();
+    }
+
     return findOne;
   }
 
@@ -27,12 +33,19 @@ export class PostsRepository {
       updatePostDto,
       { new: true },
     );
+    if (!updateOne) {
+      throw new NotFoundPostException();
+    }
 
     return updateOne;
   }
 
   async deleteOne(id: string): Promise<any> {
     const deleteOne = await this.postModel.findByIdAndDelete(id);
+    if (!deleteOne) {
+      throw new NotFoundPostException();
+    }
+
     return deleteOne;
   }
 }
